requirements.txt
absl-py==2.0.0
annotated-types==0.6.0
anyio==4.1.0
asgiref==3.7.2
astunparse==1.6.3
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
distro==1.8.0
Django==5.0
djangorestframework==3.14.0
editdistance==0.6.2
efficientnet==1.0.0
essential-generators==1.0
exceptiongroup==1.2.0
filelock==3.13.1
Flask-Cors==4.0.0
flatbuffers==23.5.26
fonttools==4.46.0
fsspec==2023.12.2
gast==0.5.4
google-api-core==2.15.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.19.4
idna==3.6
imageio==2.33.1
imgaug==0.4.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
keras==2.15.0
Keras-Applications==1.0.8
keras-ocr==0.9.3
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
ml-dtypes==0.2.0
mpmath==1.3.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
openai==1.3.8
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
Pillow==10.1.0
proto-plus==1.23.0
protobuf==4.23.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyclipper==1.3.0.post5
pydantic==2.5.2
pydantic_core==2.14.5
pyparsing==3.1.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.1
scikit-image==0.22.0
scipy==1.11.4
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0.post1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
tifffile==2023.12.9
tokenizers==0.15.0
torch==2.1.1
tqdm==4.66.1
transformers==4.36.0
triton==2.1.0
typing_extensions==4.9.0
urllib3==2.1.0
validators==0.22.0
waitress==2.1.2
Werkzeug==3.0.1
wrapt==1.14.1
